add minor changes in code namespaces
[tmp_root_ca.rb] change namespace from LogTransferIntegration to
RobyApp and deleted LogTransferIntegration module

[spawn_server.rb] includes new attr_acessors

[plugin.rb] changes reference for attributes user and password get
defined in plugin.rb
