+++ lib/syskit/roby_app/log_transfer_server/spawn_server.rb
@@ -7,7 +7,7 @@ module Syskit
         module LogTransferServer
             # Class responsible for spawning an FTP server for transfering logs
             class SpawnServer
-                attr_reader :port
+                attr_reader :port, :user, :password, :certfile_path
 
                 # tgt_dir must be an absolute path
                 def initialize(
